requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python_multipart
python-jose[cryptography]
passlib[bcrypt]
pymsql